Modal Service. Create modals from everywhere! #2155
